+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/content/component/ItemEdible.java
@@ -8,7 +8,6 @@ import house.greenhouse.bovinesandbuttercups.registry.BovinesRegistryKeys;
 import house.greenhouse.bovinesandbuttercups.util.TooltipUtil;
 import io.netty.buffer.ByteBuf;
 import net.minecraft.core.Holder;
-import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.codec.ByteBufCodecs;
@@ -16,16 +15,13 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.util.StringRepresentable;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.TooltipProvider;
 
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Function;
-import java.util.stream.Collectors;
 
 public record ItemEdible(Holder<EdibleBlockType> holder, List<MobEffectEntry> effects) implements TooltipProvider {
     public static final Codec<ItemEdible> DIRECT_CODEC = RecordCodecBuilder.create(inst -> inst.group(
@@ -44,17 +40,6 @@ public record ItemEdible(Holder<EdibleBlockType> holder, List<MobEffectEntry> ef
             ItemEdible::effects,
             ItemEdible::new
     );
-
-    public static void apply(ItemStack stack, ItemEdible edible) {
-        if (edible.holder.isBound()) {
-            var dataComponentTypes = stack.getComponentsPatch().entrySet().stream().map(Map.Entry::getKey).collect(Collectors.toSet());
-            if (!dataComponentTypes.contains(DataComponents.MAX_STACK_SIZE))
-                stack.set(DataComponents.MAX_STACK_SIZE, edible.holder.value().maxStackSize());
-            if (!dataComponentTypes.contains(DataComponents.ITEM_MODEL) && edible.holder.value().itemModel().isPresent())
-                stack.set(DataComponents.ITEM_MODEL, edible.holder.value().itemModel().get());
-            stack.set(BovinesDataComponents.EDIBLE_TYPE, edible);
-        }
-    }
 
     @Override
     public boolean equals(Object other) {